requirements.txt
polars[xlsx]>=1.0.0
pyodbc>=5.0.0
SQLAlchemy>=2.0.0
python-dotenv>=1.0.0
xlsxwriter>=3.2.0
openpyxl>=3.1.0
